Spark Structured Streaming checkpointing integration test (#459)
Changes:
- Spark version upgrade to 2.3.2
- Spark structured streaming checkpointing recovery integration test
- session-scoped shared KDC, HDFS, and Kafka services
- separate service accounts and secrets for Kafka and HDFS
- shared KDC, HDFS, and Kafka configuration and tooling was extracted to 'fixtures' modules
- Kerberized Kafka Feeder and Spark Structured Streaming test jobs

+++ testing/sdk_auth.py
@@ -195,7 +195,7 @@ class KerberosEnvironment:
                 log.info("Found installed KDC app, reusing it")
                 return _get_kdc_task(self.app_definition["id"])
             log.info("Found installed KDC app, destroying it first")
-            sdk_marathon.destroy(self.app_definition["id"])
+            sdk_marathon.destroy_app(self.app_definition["id"])
 
         log.info("Installing KDC Marathon app")
         _install_marathon_app(self.app_definition)
@@ -439,12 +439,14 @@ class KerberosEnvironment:
         sdk_security.install_enterprise_cli()
 
         log.info("Removing the marathon KDC app")
-        sdk_marathon.destroy_app(self.app_definition["id"])
+        if sdk_marathon.app_exists(self.app_definition["id"]):
+            sdk_marathon.destroy_app(self.app_definition["id"])
+            if self._temp_working_dir and isinstance(self._temp_working_dir, tempfile.TemporaryDirectory):
+                log.info("Deleting temporary working directory")
+                self._temp_working_dir.cleanup()
 
-        if self._temp_working_dir and isinstance(self._temp_working_dir, tempfile.TemporaryDirectory):
-            log.info("Deleting temporary working directory")
-            self._temp_working_dir.cleanup()
-
-        # TODO: separate secrets handling into another module
-        log.info("Deleting keytab secret")
-        sdk_security.delete_secret(self.keytab_secret_path)
+            # TODO: separate secrets handling into another module
+            log.info("Deleting keytab secret")
+            sdk_security.delete_secret(self.keytab_secret_path)
+        else:
+            log.info("KDC app doesn't exist, skipping cleanup")
